Renamed symbtrdataextractor/structure_label.py -> symbtrdataextractor/StructureLabeler.py
